feat: add implementation for integration between fe and be
